Escape backslashes when redirecting
According to the URL spec (https://url.spec.whatwg.org/), a backslash is
forbidden in a URL and must be escaped. This is especially important
because browsers implement the HTML5-defined anomaly that backslashes
should be interpreted as forward slashes.

The original behaviour allows for an open redirect in pat when ServeHTTP
is used. By escaping the backslash, we can avoid this.